Add better handling for scripts
Since we generally only want to run any of the scripts once, I removed
the concept of the `once/` directories. Instead, we now use the
`~/.dotfiles` file to keep a record of the last modified dates of all of
our scripts, only running a script if it has changed or it's new.